refactor: remove hooks (#69)
The hooks pattern seems to be outdated since we no longer doing post tx checking. The only reason to have hook is to reuse the rlp bytes, which we can do it in a trivial way -- returning it after handling the block.

This pr removes the ExecutionHook and changes the way we reusing the rlp bytes.